safari-534.53-branch/LayoutTests/platform/chromium-win/fast/dom/Range/scale-page-client-rects-expected.txt

CONSOLE MESSAGE: line 15: Uncaught TypeError: Object [object Object] has no method 'scalePageBy'
This test must be run via DumpRenderTree.
